clean-up, moved nibble to separate repo
