+++ merkle/src/lib.rs
@@ -1,4 +1,3 @@
-pub mod nibbles;
 pub mod node;
 
 mod trie;
