+++ types/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "types"
-version = "0.1.0"
+version.workspace = true
 edition.workspace = true
 authors.workspace = true
 repository.workspace = true
